+++ src/components/landing/particles/Particles.tsx
@@ -1,16 +1,42 @@
-import { Canvas } from "@react-three/fiber";
+import { useRef, useMemo } from "react";
+import { useFrame } from "@react-three/fiber";
+import { Points, PointMaterial } from "@react-three/drei";
 
-import { Experience } from "./Experience";
+import { useTheme } from "@/components/providers/ThemeProvider";
+
+const count = 200;
 
 export const Particles = () => {
+  const { darkMode } = useTheme();
+
+  const points = useRef<React.ElementRef<typeof Points>>(null);
+
+  const particlePositions = useMemo(() => {
+    const positions = new Float32Array(count * 3);
+    for (let i = 0; i < count; i++) {
+      positions[i * 3] = (Math.random() - 0.5) * 10;
+      positions[i * 3 + 1] = (Math.random() - 0.5) * 10;
+      positions[i * 3 + 2] = (Math.random() - 0.5) * 10;
+    }
+    return positions;
+  }, []);
+
+  useFrame((_state, delta) => {
+    if (points.current) {
+      points.current.rotation.x += delta * 0.01;
+      points.current.rotation.y += delta * 0.01;
+    }
+  });
+
   return (
-    <Canvas
-      camera={{ position: [0, 0, 6], fov: 50, far: 100, near: 0.1 }}
-      fallback={null}
-      gl={{ antialias: false }}
-      flat
-    >
-      <Experience />
-    </Canvas>
+    <Points ref={points} positions={particlePositions} stride={3}>
+      <PointMaterial
+        transparent
+        color={darkMode ? "#F7C19B" : "#113D61"}
+        size={0.02}
+        sizeAttenuation={true}
+        depthWrite={false}
+      />
+    </Points>
   );
 };
